backend/requirements.txt
Flask==3.0.3
flask_cors==4.0.1
protobuf==3.20.3
python-dotenv==1.0.1
requests>=2.31.0
bilibili-api-python